requirements.txt
kserve
numpy
pillow
torch
torchvision
